Reorder tests and reduce duplication
