+++ Peisik.Compiler.Tests/Compiler/Optimizing/CodeGeneratorPeisikTests.cs
@@ -2,12 +2,24 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 using Polsys.Peisik.Parser;
+using Polsys.Peisik.Compiler;
 using Polsys.Peisik.Compiler.Optimizing;
 
 namespace Polsys.Peisik.Tests.Compiler.Optimizing
 {
     class CodeGeneratorPeisikTests : CompilerTestBase
     {
+        private CompiledProgram CompileProgram(string source)
+        {
+            var syntax = ParseStringWithoutDiagnostics(source);
+            var compiler = new OptimizingCompiler(new List<ModuleSyntax>() { syntax }, Optimization.None);
+            var function = Function.FromSyntax(syntax.Functions[0], compiler);
+            var codeGen = new CodeGeneratorPeisik();
+
+            codeGen.CompileFunction(function);
+            return codeGen.Result;
+        }
+
         [Test]
         public void ConstantReturning()
         {
@@ -16,13 +28,7 @@ public int Main()
 begin
   return 5
 end";
-            var syntax = ParseStringWithoutDiagnostics(source);
-            var compiler = new OptimizingCompiler(new List<ModuleSyntax>() { syntax }, Optimization.None);
-            var function = Function.FromSyntax(syntax.Functions[0], compiler);
-            var codeGen = new CodeGeneratorPeisik();
-
-            codeGen.CompileFunction(function);
-            var program = codeGen.Result;
+            var program = CompileProgram(source);
 
             Assert.That(program, Is.Not.Null);
             var disasm = @"Int main() [0 locals]
@@ -41,13 +47,7 @@ begin
   int result 5
   return result
 end";
-            var syntax = ParseStringWithoutDiagnostics(source);
-            var compiler = new OptimizingCompiler(new List<ModuleSyntax>() { syntax }, Optimization.None);
-            var function = Function.FromSyntax(syntax.Functions[0], compiler);
-            var codeGen = new CodeGeneratorPeisik();
-
-            codeGen.CompileFunction(function);
-            var program = codeGen.Result;
+            var program = CompileProgram(source);
 
             Assert.That(program, Is.Not.Null);
             var disasm = @"Int main() [0 locals]
