+++ Peisik.Compiler.Tests/Compiler/Optimizing/LocalVariableContextTests.cs
