+++ Peisik.Compiler.Tests/Compiler/Optimizing/SequenceExpressionTests.cs
